requirements.txt
Django==6.0.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
requests==2.32.5
beautifulsoup4==4.14.3

# Local GPU summarization (MPS/CUDA capable)
torch>=2.0.0
transformers>=4.30.0
